+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -432,4 +432,9 @@ describe "The Codesake Dawn knowledge base" do
     sc.should_not   be_nil
     sc.class.should == Codesake::Dawn::Kb::NokogiriDos20131217
   end
+  it "must have test for Nokogiri_EntityExpansion_Dos_20131217" do
+    sc = kb.find("Nokogiri - Entity expasion denial of service - 20131217")
+    sc.should_not   be_nil
+    sc.class.should == Codesake::Dawn::Kb::Nokogiri_EntityExpansion_Dos_20131217
+  end
 end